refactor: combine statement and parser files
